FEAT-OPTIMISTIC Optimistic Eth block building based on notarization hint.
Eth/Geth cannot build a block until its parent block is known to the canonical
chain.

The earliest opportunity to build a block comes after first round of consensus
where the block consensus is not finalized.

To build a block early the notarized block has to be
"tentatively/optimistically" be set as head before triggering block building.